+++ lib/libc/string/memset.c
@@ -16,7 +16,7 @@ void * memset(void * m, int c, size_t n)
 
 /* If optmizing for speed */
 #ifndef __OPTIMIZE_SIZE__
-    int i;
+    size_t i;
     unsigned long buffer;
     unsigned long *aligned_addr;
     unsigned int d = c & 0xff;      /* To avoid sign extension, copy C to an
@@ -43,13 +43,13 @@ void * memset(void * m, int c, size_t n)
             buffer = (buffer << i) | buffer;
 
         /* Unroll the loop. */
-        while (n >= LBLOCKSIZE*4)
+        while (n >= LBLOCKSIZE * 4)
         {
             *aligned_addr++ = buffer;
             *aligned_addr++ = buffer;
             *aligned_addr++ = buffer;
             *aligned_addr++ = buffer;
-            n -= 4*LBLOCKSIZE;
+            n -= 4 * LBLOCKSIZE;
         }
 
         while (n >= LBLOCKSIZE)
@@ -58,7 +58,7 @@ void * memset(void * m, int c, size_t n)
             n -= LBLOCKSIZE;
         }
         /* Pick up the remainder with a bytewise loop. */
-        s = (char*)aligned_addr;
+        s = (char *)aligned_addr;
     }
 #endif
